fix: move powered by tag to input and fix padding on full chat
